Edit:102

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,14 +5,14 @@
  */
 int main (void)
 {
-	int a = 1, b = 2, n = 2, d;
+	long int a = 1, b = 2, n = 2, d;
 
-	printf("%i, %i", a, b);
+	printf("%lu, %lu", a, b);
 
 	while (n < 50)
 	{
 		d = a + b;
-		printf(", %i", d);
+		printf(", %lu", d);
 		a = b;
 		b = d;
 		n++;
